requirements.txt
blueprint==3.4.2
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pip==19.0.3
pyMySql==0.9.3
PyJWT==1.7.1
python-dotenv==0.10.1
PyYAML==3.13
setuptools==40.8.0
six==1.12.0
Werkzeug==0.14.1
Sphinx==1.8.4
